requirements.txt
anthropic>=0.92.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
uvicorn[standard]
